move user to knn package
